Fix CI/CD: Comprehensive sign conversion fixes for Clang-15
Fix all remaining -Wsign-conversion errors that only appear with Clang-15:

**funcstate.cpp:**
- Line 451: Cast offset to unsigned int in SETARG_Bx

**lcode.cpp (10 fixes):**
- Line 147: Cast reg to unsigned int in SETARG_A (patchtestreg)
- Line 537: Cast reg to unsigned int in SETARG_A (discharge2reg)
- Line 922: Cast int2sC() result to unsigned int in SETARG_B
- Lines 1058-1059: Cast getInfo() and n+1 to unsigned int in SETARG_A/B
- Lines 1171-1172: Cast from and l-from to unsigned int in SETARG_A/B
- Lines 1423,1426-1427: Cast nresults+1 and getFreeReg() to unsigned int
- Line 1674: Cast getNumParams()+1 to unsigned int in SETARG_C

**parser.cpp (1 fix):**
- Line 630: Cast cc.na and cc.nh to unsigned in settablesize() call

**parseutils.cpp (1 fix):**
- Lines 161-178: Remove unused priority[] array (already defined in parser.cpp)

All SETARG_* macros expect unsigned int parameters but were receiving
int values, causing implicit sign conversion warnings in Clang-15.
GCC is more permissive and doesn't catch these.

Tests: All passing, performance unchanged at 2.24s

+++ src/compiler/funcstate.cpp
@@ -448,7 +448,7 @@ void FuncState::fixforjump(int pcpos, int dest, int back) {
     offset = -offset;
   if (l_unlikely(offset > MAXARG_Bx))
     getLexState()->syntaxError("control structure too long");
-  SETARG_Bx(*jmp, offset);
+  SETARG_Bx(*jmp, static_cast<unsigned int>(offset));
 }
